requirements.txt
pandas==1.4.2
requests==2.27.1
bs4==0.0.1
lxml==4.9.1
python-telegram-bot==13.14
